поправил шапку в зависимости от LoginStatus
